Always define sTHX as 0 when not using mem debug header
I broke the threaded build in b001a0d149 by assuming that sTHX was
always set to 0 when memory debug headers were not in use.  That is
an easy assumption to make, so this commit makes that assumption true.
(And there is no reason it cannot be made true.  There is no need for
sTHX to be set under PERL_IMPLICIT_CONTEXT but not PERL_TRACK_MEMPOOL,
where sTHX was not even used before b001a0d149.)