pod/perllexwarn.pod: clarify note about the "deprecated" category
Saying that something used to behave a certain way "In Perl 5.6.1" can
be understood as it changing in 5.6.2, but in fact this change to the
"deprecated" category in perl-5.6.0-8156-g12bcd1a was never released in
any 5.6.* release, it was first released in 5.8.0.

So note that explicitly instead, it's less ambiguous to just say "Before
Perl 5.8.0".