[perl #120374] Stop for($h{k}||'') from vivifying
Commit 2e73d70e52 broke this (made it vivify) by propagating lvalue
context to the branches of || and && (to fix another bug).  It broke
App::JobLog as a result.

Because foreach does not do defelem magic (i.e., it vivifies), this
ends up extending vivification to happen where it did not before.

Fixing foreach to do defelem magic (create ‘deferred element’ scalars,
the way sub calls do, to avoid vivifying immediately) would be another
way to fix this.  (And I plan to do it soon anyway.  That is the topic
of ticket #115194.)

However, to create a deferred element on the lhs of || would be a
waste, since the newly-created SV will evaluate to false, and will be
thrown away immediately.

So, when the lhs of || is helem or aelem, we can simply refram from
applying lvalue context, and the result is exactly the same as we
would get from applying defelem magic, but faster.