Allow 'use locale' on systems without locales.
Instead of throwing an error, just go ahead and do the import.
This will tell Perl internally to use the current underlying locale,
which should be the C locale.  Attempts to change the locale will fail.

This differs slightly from Brian Fraser's patch, in that his didn't
touch $^H, thus 'use locale' was a no-op.  He has told me to apply this
one, which does affect $^H.  The advantage here is that now programs
that are run on platforms with and without locales will behave
similarly, and should run identically if the locale is not switched from
the default.