re_intuit_start(): rearrange /^/m code
After matching the "check" and "other" strings, we check that
rx_origin is at a \n in the presence of /^../m. The code that
does this is in one half of an if-statement, with a couple of labels and
gotos that get us to and from the other half of the if statement.

Re-arrange the code so that the /^../m is done on its own before the if.
This removes a couple of labels and gotos and makes the code clearer.

Basically we went from:

    if (rx_origin != strpos) {
        if (ml_anch && COND_A) {
          find_anchor:
            LOOK_FOR_ANCHOR...
        }
        REST_A;
    }
    else {
        if (ml_anch && COND_B) {
            goto find_anchor;
        }
        REST_B;
    }

to:

    if (rx_origin != strpos && (ml_anch && COND_A)
    ||  rx_origin == strpos && (ml_anch && COND_B))
    {
          find_anchor:
            LOOK_FOR_ANCHOR...
        }
        ...
    }

    if (rx_origin != strpos) {
        REST_A;
    else {
        REST_B;
    }

The next couple of commits will re-indent and simplify the condition a
bit.