re_intuit_start(): use memchr() to find \n
The code that scans for another \n when the /^.../m constraint has failed
currently searches for the \n in a big while loop. Use memchr() instead,
which is likely to to be as efficient as possible on any given platform.
Also invert the sense of most of the tests following a found \n,
which removes lots indented ifs, and leaves us with a clean set of

    if (A)
        goto X;
    if (B)
        goto Y;
    if (C)
        goto Z;