re_intuit_start(): make assert unconditional
an assert was originally in the float-then-anchored branch, but not in the
anchored-then-float branch. When the two branches were merged, the assert
was only done if other==anchored. It turns out the the assert should be
true in both cases, so remove the guard.

I've also changed the condition from

    prog->minlen > other->min_offset
to
    prog->minlen >= other->min_offset

Since they can in fact be equal on a one-char substr with SvTAIL().