re_intuit_start(): don't decrease other_last
The /^../m failure code did an unconditional other_last = rx_origin;
if other_last was already high, it could get shrunk and we'd end
up running fbm over the same bit of string repeatedly.

The following code

    $s = "-ab\n" x 500_000;
    $s .= 'abx';
    $s =~ /^ab.*x/m;

(which went quadratic on length) reduces from minutes to millisecs with
this commit.  This is because we'd keep going back to near the beginning
of the string and searching for 'x' again.