Expand tabs in diagnostics.pm
Otherwise pod like this:

The second situation is caused by an eval accessing a lexical subroutine
that has gone out of scope, for example,

    sub f {
	my sub a {...}
	sub { eval '\&a' }
    }
    f()->();

is turned into this:

    The second situation is caused by an eval accessing a variable that has
    gone out of scope, for example,

        sub f {
    	my $a;
    	sub { eval '$a' }
        }
        f()->();

instead of this:

    The second situation is caused by an eval accessing a variable that has
    gone out of scope, for example,

        sub f {
            my $a;
            sub { eval '$a' }
        }
        f()->();

I don’t know how to test this without literally copying and pasting
parts of diagnostics.pm into diagnostics.t.  But I have tested it man-
ually and it works.