pp_concat: Only call SvPV_force_nolen when needed
If we just did an sv_setpvs on it, the SvPV_force_nolen should not do
anything useful, so let's not.

Side note: s/TARG/left/ in the enclosing block because they are the same
pointer, so why use a define that needs grokking by the reader of the
code if the local variable is guaranteed to be the same?